// FE/FE_Itsol/src/app/@core/models/company.ts
export class Company {
  id: number;
  name: string;
  email: string;
  hotline: string;
  date_incoporation: Date;
  tax_code: string;
  tax_date: Date;
  tax_place: string;
  head_office: string;
  number_staff: number;
  link_web: string;
  description: string;
  avatar: string;
  backdrop_img: string;
  is_delete: boolean;
}